update(server): :jigsaw: update
update WorkerManager

+++ server/src/utils/WorkerManager.ts
@@ -1,21 +1,17 @@
 import WorkerPool from 'workerpool'
 import Pool from 'workerpool/src/Pool'
-import { PROCESS_ENV } from './InitEnv'
 import Console from './ConsoleHandler'
 
 interface IInitOptions {
 	minWorkers: number
 	maxWorkers: number
+	workerTerminateTimeout?: number
 }
 
 interface IGetFreePool {
 	pool: Pool
 	terminate: () => void
 }
-
-const MAX_WORKERS = PROCESS_ENV.MAX_WORKERS
-	? Number(PROCESS_ENV.MAX_WORKERS)
-	: 7
 
 const WorkerManager = (() => {
 	return {
@@ -26,9 +22,12 @@ const WorkerManager = (() => {
 		) => {
 			options = {
 				minWorkers: 1,
-				maxWorkers: MAX_WORKERS,
+				maxWorkers: 1,
+				workerTerminateTimeout: 0,
 				...(options || {}),
 			}
+
+			const MAX_WORKERS = options.maxWorkers
 
 			let curPool = WorkerPool.pool(workerPath, options)
 
@@ -45,43 +44,58 @@ const WorkerManager = (() => {
 				Console.error(err)
 			}
 
-			const _terminate = (() => {
-				let isTerminate = false
+			const _terminate = (pool: Pool) => {
+				let count = 0
+				// let timeout: NodeJS.Timeout
+				return () => {
+					count++
 
-				return async () => {
-					if (isTerminate) return
-					isTerminate = true
-					const pool = curPool
+					if (count === MAX_WORKERS) {
+						setTimeout(() => {
+							pool.terminate()
+						}, 5000)
+					}
+				}
+			}
 
-					const newPool = WorkerPool.pool(workerPath, {
-						...options,
-					})
+			const _getFreePool: () => IGetFreePool = (() => {
+				let count = 0
+				let pool = curPool as Pool
+				let terminate = _terminate(pool)
 
-					try {
-						if (instanceTaskList && instanceTaskList.length) {
-							const promiseTaskList: Promise<any>[] = []
-							for (const task of instanceTaskList) {
-								promiseTaskList.push(newPool.exec(task, []))
-							}
-
-							await Promise.all(promiseTaskList)
-						}
-					} catch (err) {
-						Console.error(err)
-					} finally {
-						isTerminate = false
+				return () => {
+					count++
+					if (count > MAX_WORKERS) {
+						count = 1
+						pool = curPool
+						terminate = _terminate(pool)
 					}
 
-					curPool = newPool
-					isTerminate = false
-					pool.terminate()
-				}
-			})()
+					if (count === 1) {
+						curPool = WorkerPool.pool(workerPath, {
+							...options,
+						})
 
-			const _getFreePool: () => IGetFreePool = () => ({
-				pool: curPool,
-				terminate: _terminate,
-			}) // _getFreePool
+						try {
+							if (instanceTaskList && instanceTaskList.length) {
+								const promiseTaskList: Promise<any>[] = []
+								for (const task of instanceTaskList) {
+									promiseTaskList.push(curPool.exec(task, []))
+								}
+
+								Promise.all(promiseTaskList)
+							}
+						} catch (err) {
+							Console.error(err)
+						}
+					}
+
+					return {
+						pool,
+						terminate,
+					}
+				}
+			})() // _getFreePool
 
 			return {
 				getFreePool: _getFreePool,
